+++ corpus/schema/defs.clj
@@ -29,3 +29,35 @@
 
 (schema.core/defn baz []
   (str {:a 1 :b}))
+
+(s/defn bad-return1 []
+  :- s/Int ;; Return schema should go before vector.
+  1)
+
+(s/defn bad-return2
+  ([]
+   :- s/Int ;; Return schema should go before arities.
+   1))
+
+(s/defn bad-return3
+  (:- s/Int [] ;; Function arguments should be wrapped in vector.
+   1))
+
+(s/defn :- s/Int ;; TODO not detected by clj-kondo but immediately fails expansion when evaluated
+  bad-return4
+  [bad4]
+  1)
+
+(s/defn bad-return5
+  "foo"
+  :- s/Int ;; TODO not detected by clj-kondo but immediately fails expansion when evaluated
+  []
+  1)
+
+(s/defn bad-defn
+  ()) ;;Invalid function body.
+
+(s/defn good-return1 [] :-)
+(s/defn good-return2 ([] :-))
+
+#_:clj-kondo/ignore (s/defn foo [] :- s/Str) ;; schema-misplaced-return can be ignored